form submission redirects to newChapter#submitted
If the form redirects to the homepage (as it did previously), then
the flask homepage code needs to be able to handle the form submission.
That's ugly code, so I'm keeping the form submission handling in the
newChapter method.

A neat fix might be to set up a redirect to the homepage through
either flask, or the newChapter#submitted url. I'm not really sure
what would be best though